api/worker/requirements.txt
# Worker service dependencies
# Core framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.7.4
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis and event streaming
redis==5.0.1
aioredis==2.0.1

# AI providers
openai>=1.50.0
langchain>=0.3.0
langchain-gigachat==0.3.12
langchain-core>=0.3.0
langgraph>=0.2.3

# Vector database
qdrant-client==1.15.1

# Graph database
neo4j>=5.25.0,<6.0.0

# Web crawling and enrichment
# Context7: Убрали crawl4ai и playwright - используем внешний crawl4ai сервис через Redis Streams
beautifulsoup4==4.12.2
lxml>=5.3.0

# HTTP client
httpx>=0.27.2
aiohttp>=3.11.11

# S3 Storage (Cloud.ru) - Context7: для S3 lifecycle и storage quota
boto3>=1.34.0
botocore>=1.34.0

# Image processing utilities
imagehash==4.3.1
numpy>=2.0.0,<3.0.0

# Monitoring and metrics
prometheus-client==0.19.0
structlog==23.2.0

# Configuration
python-dotenv==1.0.0
pyyaml>=6.0.3,<7.0.0

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development and testing
pytest==7.4.4
pytest-asyncio==0.21.2
pytest-mock==3.12.0
jsonschema>=4.21.1

# Logging and debugging
rich>=13.9.4
loguru==0.7.2

mistune>=3.0.0

aiogram>=3.13.0

# OCR Enhancement (опциональные зависимости)
# Context7: pyspellchecker и langdetect опциональны - код работает без них
pyspellchecker>=0.8.0; python_version >= "3.11"
langdetect>=1.0.9; python_version >= "3.11"